Improvements and rework of internals

+++ src/NodeData/GNSS/GnssNavInfo.hpp
@@ -106,7 +106,8 @@ class GnssNavInfo
                 std::string ret;
                 for (const auto& navData : m_satellites.at(satId).getNavigationData())
                 {
-                    ret += fmt::format("[{} - diff {:.0f}s], ", navData->refTime.toYMDHMS(GPST), std::abs((navData->refTime - recvTime).count()));
+                    ret += fmt::format("[{} - diff {:.0f}s], ", navData->refTime.toYMDHMS(GPST),
+                                       static_cast<double>(std::abs((navData->refTime - recvTime).count())));
                 }
                 return ret.substr(0, ret.length() - 2);
             };
