+++ src/Nodes/DataProvider/IMU/Sensors/VectorNavSensor.cpp
@@ -7415,7 +7415,7 @@ void NAV::VectorNavSensor::asciiOrBinaryAsyncMessageReceived(void* userData, vn:
                         if (obs->insTime - vnSensor->_lastMessageTime.at(b) >= std::chrono::duration<double>(1.5 * (vnSensor->_binaryOutputRegister.at(b).rateDivisor / IMU_DEFAULT_FREQUENCY)))
                         {
                             LOG_WARN("{}: Potentially lost a message. dt = {:.4} s, expect {} s. (Previous message at [{}], current message [{}])", vnSensor->nameId(),
-                                     (obs->insTime - vnSensor->_lastMessageTime.at(b)).count(),
+                                     static_cast<double>((obs->insTime - vnSensor->_lastMessageTime.at(b)).count()),
                                      1. / IMU_DEFAULT_FREQUENCY * vnSensor->_binaryOutputRegister.at(b).rateDivisor,
                                      vnSensor->_lastMessageTime.at(b), obs->insTime);
                         }
